pyproject.toml
[project]
name = "otxserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
